requirements.txt
# FastAPI相关
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# 数据库相关
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Redis相关
redis==5.0.1

# AI和机器学习相关
openai==1.3.7
langchain==0.0.350
langchain-core==0.1.3
langgraph==0.0.26

# 数据处理
pandas==2.1.4
numpy==1.24.4
pydantic==2.5.1

# HTTP客户端
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.1

# 文件处理
python-dotenv==1.0.0
PyYAML==6.0.1

# 系统监控
psutil==5.9.6

# 其他工具
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2

# 可选：如果需要语音功能
# speechrecognition==3.10.0
# pyaudio==0.2.11

# 可选：如果需要图像处理
# Pillow==10.1.0
# opencv-python==4.8.1.78 